some refactoring for bomb creation
